detect 10695896

+++ code/10693040.py
@@ -3,15 +3,16 @@ import os
 import matplotlib.pyplot as plt
 import easyocr
 import shutil
+import json
+import re
 
 
 data_dict = {
-    # id, sentences per page, wide_crop, margin_x, width_ths_pass
-    10693040: (12, 400, 250, 10),      
+    # id: sentences per page, wide_crop, margin_x, margin_y, width_ths_pass, json, width_ths_word_2, first_word_distance, last_word_distance
+    10693040: (12, 400, 250, 120, 10, 'data/BichCauKyNgo.json', 0.1, 90, 150),      
     10693454: (11, 400, 200, 10),       
     10695896: (12, 200, 290, 10),     
-    10696073: (10, 400),       
-    10709453: (15, 7)       
+    10723635: (12, 550, 230, 100, 10, 'data/TrinhThu.json', 0.3, 90, 150),            
 }
 
 def shift_and_insert_image(folder_path, new_image_path, insert_index):
@@ -32,27 +33,33 @@ def shift_and_insert_image(folder_path, new_image_path, insert_index):
     new_image_name = f"cropped_word_{insert_index:03}.png"
     shutil.copy(new_image_path, os.path.join(folder_path, new_image_name))
 
+def parse_json(json_file):
+    with open(json_file, 'r') as f:
+        json_data = f.read()
+        parsed_json = json.loads(json_data)
+
+    # Extracting all values from each dictionary in the list
+    result = []
+    for item in parsed_json:
+        for key, value in item.items():
+            result.extend(value)  
+    return result
+
 # Path to the folder containing images
-input_folder = 'processed_images/10693454'  
+input_folder = 'processed_images/10693040'  
 pdf_id = os.path.basename(input_folder) 
 output_folder = f'image_crop/{pdf_id}'
+
 sentence_per_image = data_dict[int(pdf_id)][0]
 crop_image_wide = data_dict[int(pdf_id)][1]
 margin_x = data_dict[int(pdf_id)][2]
-width_ths_pass = data_dict[int(pdf_id)][3]
-def delete_box(boxes):
-    h = len(boxes)
-    i = 0
-    while i < h:
-        if abs(boxes[i][0][0][1] - boxes[i][0][2][1]) >= 150:
-            boxes.pop(i)
-            h -= 1
-        elif abs(boxes[i][0][0][0] - boxes[i][0][1][0]) <= 400:
-            boxes.pop(i)
-            h -= 1
-        else:
-            i += 1
-    return boxes
+margin_y = data_dict[int(pdf_id)][3]
+width_ths_pass = data_dict[int(pdf_id)][4]
+json_file = parse_json(data_dict[int(pdf_id)][5])
+width_ths_word_2 = data_dict[int(pdf_id)][6]
+first_word_distance = data_dict[int(pdf_id)][7]
+last_word_distance = data_dict[int(pdf_id)][8]
+current_sentence = 0
 
 def is_box_within(merged_box, non_merged_box):
     top_left_merged = merged_box[0]
@@ -76,8 +83,10 @@ if not os.path.exists(output_folder):
 # Initialize the OCR reader
 reader = easyocr.Reader(['vi', 'en'])  
 
+directories = [d for d in os.listdir(input_folder)]
+directories.sort()
 # Iterate through each image in the input folder
-for filename in os.listdir(input_folder):
+for filename in directories:
     if filename.lower().endswith(('.png', '.jpg', '.jpeg')): 
         image_path = os.path.join(input_folder, filename)
         image = cv2.imread(image_path)
@@ -90,23 +99,12 @@ for filename in os.listdir(input_folder):
         # Extract bounding boxes and sort by the top-left y coordinate
         boxes = [(result[0], result[1]) for result in results]  # List of (bounding box, text)
         boxes.sort(key=lambda x: x[0][0][1])  # Sort by the y coordinate of the top-left corner
-        #Draw bounding box without adding text
-        # for box in boxes:
-        #     (top_left, top_right, bottom_right, bottom_left) = box[0]
-        #     top_left = tuple(map(int, top_left))
-        #     bottom_right = tuple(map(int, bottom_right))
-        #     cv2.rectangle(image, top_left, bottom_right, (0, 255, 0), 2)  # Bounding box in green
-        # plt.imshow(cv2.cvtColor(image, cv2.COLOR_BGR2RGB))
-        # plt.axis('off')  # Hide axes
-        # plt.show()
-
-        #boxes = delete_box(boxes)
 
         # Group boxes and merge as needed (logic as provided in previous step)
         merged_boxes = []
         non_merged_boxes = []  # List for non-merged boxes
         current_group = []
-        variance_threshold = 40  # Y-coordinate variance threshold
+        variance_threshold = 50  # Y-coordinate variance threshold
 
         for bbox, text in boxes:
             if not current_group:
@@ -163,19 +161,18 @@ for filename in os.listdir(input_folder):
             # Get the top-left and bottom-right points of the box
             top_left = tuple(map(int, box[0]))
             bottom_right = tuple(map(int, box[2]))
-            if bottom_right[1] <= 200:
+            if bottom_right[1] <= margin_y:
                 continue
+
+            if bottom_right[0] <= margin_x:
+                continue 
 
             if top_left[0] < margin_x:
                 top_left = (margin_x, top_left[1])
 
-            if abs(bottom_right[0]) >= 400:
-                cropped_image = image[top_left[1]-10:bottom_right[1]+10, max(top_left[0]-100, margin_x):min(bottom_right[0]+200, image.shape[1])]
-            else:
-                continue
+            cropped_image = image[top_left[1]-10:bottom_right[1]+10, max(top_left[0]-130, margin_x):min(bottom_right[0]+200, image.shape[1])]
 
-            if cropped_image is None or cropped_image.size == 0 or abs(cropped_image.shape[1]) <= crop_image_wide: # x<=400
-                # print("Empty image, skipping...")
+            if cropped_image is None or cropped_image.size == 0 or abs(cropped_image.shape[1]) <= crop_image_wide: 
                 continue
             else:
                 index = index + 1                
@@ -189,11 +186,17 @@ for filename in os.listdir(input_folder):
                 new_file = os.path.join(cropped_image_folder, f'cropped_word')
                 if not os.path.exists(new_file):
                     os.makedirs(new_file)                
-                results_word = reader.readtext(cropped_image_path, height_ths =1.5, slope_ths = 0.9, width_ths = 0.1) 
+                results_word = reader.readtext(cropped_image_path, height_ths =1.5, slope_ths = 5, width_ths = 0.1) 
                 word = [(result[0], result[1]) for result in results_word]  # List of (bounding box, text)
                 word.sort(key=lambda x: x[0][0][0]) 
+                if not word:
+                    if os.path.exists(cropped_image_folder):
+                        shutil.rmtree(cropped_image_folder)
+                    index = index - 1
+                    continue
 
                 i=0
+                current_sentence = current_sentence + 1
                 while i < len(word) - 1:
                     # Calculate the widths of the current box and the next box
                     width_i = word[i][0][2][0] - word[i][0][0][0]
@@ -237,7 +240,7 @@ for filename in os.listdir(input_folder):
                     # print(filename, ' ', cropped_word_image.shape[1], ' ', cnt)
                     if cropped_word_image.shape[1] >= 200:
                         # Perform OCR on the oversized crop
-                        tmps = reader.readtext(cropped_word_image, height_ths=1, slope_ths=5, width_ths=0.01)
+                        tmps = reader.readtext(cropped_word_image, height_ths=1, slope_ths=5, width_ths=width_ths_word_2)
                         # If OCR detects multiple results, save each as a separate image
                         if len(tmps) >= 2:
                             w = [(tmp[0], tmp[1]) for tmp in tmps]  # List of (bounding box, text)
@@ -258,34 +261,15 @@ for filename in os.listdir(input_folder):
                     cropped_word_image_path = os.path.join(new_file, f'cropped_word_{cnt:03}.png')
                     cv2.imwrite(cropped_word_image_path, cropped_word_image)  
                 
-                
-                if (index % 2 == 0 and cnt != 8) or (index % 2 != 0 and cnt != 6):
+                word_count_ground = len(re.split(r'[ -]+', json_file[current_sentence - 1]))
+                if cnt != word_count_ground:
                     # Load the cropped image
                     cropped_image_1 = cv2.imread(cropped_image_path)
-                    cnt = 0
+                    tmp_cnt = 0
                     distance = 0
                     max_indices = None
 
-                    # First word
-                    if abs(word[0][0][0][0]) > 150:
-                        x_start = 0
-                        x_end = min(word[0][0][0][0], cropped_image_1.shape[1])
-                        y_start = 0
-                        y_end = cropped_image_1.shape[0]
-                        # Crop the space between the two boxes
-                        cropped_space = cropped_image_1[y_start:y_end, x_start:x_end]
-
-                        # Save the cropped space
-                        if cropped_space is not None and cropped_space.size > 0:
-                            cnt = cnt + 1
-                            cropped_space_temp_path = os.path.join(new_file, f'temp_cropped_space_{cnt:03}.png')
-                            cv2.imwrite(cropped_space_temp_path, cropped_space)
-                            # Insert the image at the correct position in the folder
-                            shift_and_insert_image(new_file, cropped_space_temp_path, 1)
-                            os.remove(cropped_space_temp_path)  # Clean up temporary file
-
-                        
-                    # Middle words
+                    # Middel words
                     for j in range(1, len(word)):
                         # Calculate the horizontal distance between top-left of current box and top-right of previous box
                         current_distance = word[j][0][0][0] - word[j-1][0][1][0]
@@ -301,47 +285,64 @@ for filename in os.listdir(input_folder):
                             # Save the cropped space
                             if cropped_space is not None and cropped_space.size > 0:
                                 cnt = cnt + 1
-                                cropped_space_temp_path = os.path.join(new_file, f'temp_cropped_space_{cnt:03}.png')
+                                tmp_cnt = tmp_cnt + 1
+                                cropped_space_temp_path = os.path.join(new_file, f'temp_cropped_space_{tmp_cnt:03}.png')
                                 cv2.imwrite(cropped_space_temp_path, cropped_space)
                                 # print(f"Temporary cropped space image saved at {cropped_space_temp_path}")
 
                                 # Insert the image at the correct position in the folder
-                                shift_and_insert_image(new_file, cropped_space_temp_path, j)
+                                shift_and_insert_image(new_file, cropped_space_temp_path, j+1)
+                                os.remove(cropped_space_temp_path)  # Clean up temporary file
+                    
+                    # Last, First word
+                    if cnt != word_count_ground:
+                        if word and abs(word[0][0][0][0]) > first_word_distance:
+                            x_start = 0
+                            x_end = int(word[0][0][0][0])
+                            y_start = 0
+                            y_end = int(cropped_image_1.shape[0])
+                            # Crop the space between the two boxes
+                            cropped_space = cropped_image_1[y_start:y_end, x_start:x_end]
+
+                            # Save the cropped space
+                            if cropped_space is not None and cropped_space.size > 0:
+                                cnt = cnt + 1
+                                tmp_cnt = tmp_cnt + 1
+                                cropped_space_temp_path = os.path.join(new_file, f'temp_cropped_space_{tmp_cnt:03}.png')
+                                cv2.imwrite(cropped_space_temp_path, cropped_space)
+                                # Insert the image at the correct position in the folder
+                                shift_and_insert_image(new_file, cropped_space_temp_path, 1)
                                 os.remove(cropped_space_temp_path)  # Clean up temporary file
 
+                        # Last word
+                        if word and abs(word[len(word)-1][0][1][0] - cropped_image_1.shape[1]) > last_word_distance:
+                            x_start = int(word[len(word)-1][0][1][0])
+                            x_end = int(cropped_image_1.shape[1])
+                            y_start = 0
+                            y_end = int(cropped_image_1.shape[0])
+                            # Crop the space between the two boxes
+                            cropped_space = cropped_image_1[y_start:y_end, x_start:x_end]
 
-                    # Last word
-                    if abs(word[len(word)-1][0][1][0] - cropped_image_1.shape[1]) > 130:
-                        x_start = word[len(word)-1][0][1][0]+10
-                        x_end = cropped_image_1.shape[1]
-                        y_start = 0
-                        y_end = cropped_image_1.shape[0]
-                        # Crop the space between the two boxes
-                        cropped_space = cropped_image_1[y_start:y_end, x_start:x_end]
+                            # Save the cropped space
+                            if cropped_space is not None and cropped_space.size > 0:
+                                cnt = cnt + 1
+                                tmp_cnt = tmp_cnt + 1
+                                cropped_space_temp_path = os.path.join(new_file, f'temp_cropped_space_{tmp_cnt:03}.png')
+                                cv2.imwrite(cropped_space_temp_path, cropped_space)
+                                # print(f"Temporary cropped space image saved at {cropped_space_temp_path}")
+                                file_count = len([f for f in os.listdir(new_file) if os.path.isfile(os.path.join(new_file, f))])
+                                shift_and_insert_image(new_file, cropped_space_temp_path, file_count)
+                                os.remove(cropped_space_temp_path)  # Clean up temporary file
+                        
+                if cnt != word_count_ground:
+                    print(new_file, ' ', cnt, ' ', word_count_ground, ' ', json_file[current_sentence-1]) #Error file
 
-                        # Save the cropped space
-                        if cropped_space is not None and cropped_space.size > 0:
-                            cnt = cnt + 1
-                            cropped_space_temp_path = os.path.join(new_file, f'temp_cropped_space_{cnt:03}.png')
-                            cv2.imwrite(cropped_space_temp_path, cropped_space)
-                            # print(f"Temporary cropped space image saved at {cropped_space_temp_path}")
-
-                            # Insert the image at the correct position in the folder
-                            if (index % 2 == 0):
-                                shift_and_insert_image(new_file, cropped_space_temp_path, 8)
-                            else:
-                                shift_and_insert_image(new_file, cropped_space_temp_path, 6)
-                            os.remove(cropped_space_temp_path)  # Clean up temporary file
-                    
-                    file_count = len([f for f in os.listdir(new_file) if os.path.isfile(os.path.join(new_file, f))])
-                    if index % 2 == 0 and file_count != 8:
-                        print(f'Error: {filename} - {index} - {file_count}') 
-                    elif index % 2 != 0 and file_count != 6:
-                        print(f'Error: {filename} - {index} - {file_count}')
-
-        if (index != sentence_per_image): print(file_output_folder) 
-
+                cropped_image_text_path = os.path.join(cropped_image_folder, f'cropped_sentence_text_{index:03}.txt')
+                text_to_store = json_file[current_sentence-1] 
+                with open(cropped_image_text_path, 'w', encoding='utf-8') as txt_file:
+                    txt_file.write(text_to_store)
+                # print(filename, ' ', index, ' ', cnt, ' ', word_count_ground, ' ', json_file[current_sentence-1])
+        if (index != sentence_per_image): 
+            current_sentence = current_sentence - (index - sentence_per_image)
+            print(file_output_folder) 
 print("All images have been processed and cropped images have been saved successfully.")
-
-
-
